Fix base cost calculation

+++ src/model/redeem.go
@@ -25,6 +25,6 @@ type ShippingCost struct {
 	gorm.Model
 	CountryId   uint    `json:"country_id"`
 	ProvinceId  uint    `json:"province_id"`
-	BaseCost    float32 `json:"base_cost"`
-	CostPerUnit float32 `json:"cost_per_unit"`
+	BaseCost    float64 `json:"base_cost"`
+	CostPerUnit float64 `json:"cost_per_unit"`
 }
